[HUB] docker file
