+++ imports/rocketchat-hub/index.js
@@ -1,84 +1,10 @@
-// import { Streamer } from './Streamer';
-// import { STREAM_NAMES } from './constants';
-const normalize = {
-	update: 'updated',
-	insert: 'inserted',
-	remove: 'removed',
-};
+import hub from './src';
+import { RocketChat } from 'meteor/rocketchat:lib';
 
-export default ({ Messages, Subscriptions, Rooms, Settings, Trash }) => ({
-	name: 'hub',
-	created() {
-		const RocketChat = {
-			Services: this.broker,
-		};
-
-		Messages.watch([], { fullDocument: 'updateLookup' }).on('change', async function({ operationType, /* documentKey,*/ fullDocument/* , oplog*/ }) {
-			switch (operationType) {
-				case 'insert':
-				case 'update':
-					// const message = await Messages.findOne(documentKey);
-					const message = fullDocument;
-					// Streamer.emitWithoutBroadcast('__my_messages__', message, {});
-					RocketChat.Services.broadcast('message', { action: normalize[operationType], message });
-						// return Streamer.broadcast({ stream: STREA	M_NAMES['room-messages'], eventName: message.rid, args: message });
-						// publishMessage(operationType, message);
-			}
-		});
-		Subscriptions.watch([], { fullDocument: 'updateLookup' }).on('change', async({ operationType, documentKey, fullDocument }) => {
-			let subscription;
-			switch (operationType) {
-				case 'insert':
-				case 'update':
-					// subscription = await Subscriptions.findOne(documentKey/* , { fields }*/);
-					subscription = fullDocument;
-					break;
-
-				case 'remove':
-					subscription = await Trash.findOne(documentKey, { fields: { u: 1, rid: 1 } });
-					break;
-				default:
-					return;
-			}
-
-			RocketChat.Services.broadcast('subscription', { action: normalize[operationType], subscription });
-		});
-		Rooms.watch([], { fullDocument: 'updateLookup' }).on('change', async({ operationType, documentKey, fullDocument }) => {
-			let room;
-			switch (operationType) {
-				case 'insert':
-				case 'update':
-					// room = await Rooms.findOne(documentKey/* , { fields }*/);
-					room = fullDocument;
-					break;
-
-				case 'remove':
-					room = documentKey;
-					break;
-				default:
-					return;
-			}
-			// console.log(room, documentKey);
-			RocketChat.Services.broadcast('room', { action: normalize[operationType], room });
-			// RocketChat.Notifications.streamUser.__emit(data._id, operationType, data);
-		});
-		Settings.watch([], { fullDocument: 'updateLookup' }).on('change', async({ operationType, documentKey, fullDocument }) => {
-			let setting;
-			switch (operationType) {
-				case 'insert':
-				case 'update':
-					// setting = Settings.findOne(documentKey/* , { fields }*/);
-					setting = fullDocument;
-					break;
-				case 'remove':
-					setting = documentKey;
-					break;
-				default:
-					return;
-			}
-
-			RocketChat.Services.broadcast('setting', { action: normalize[operationType], setting });
-			// RocketChat.Notifications.streamUser.__emit(data._id, operationType, data);
-		});
-	},
+export default hub({
+	Trash: RocketChat.models.Trash.rawCollection(),
+	Messages: RocketChat.models.Messages.model.rawCollection(),
+	Subscriptions: RocketChat.models.Subscriptions.model.rawCollection(),
+	Rooms: RocketChat.models.Rooms.model.rawCollection(),
+	Settings: RocketChat.models.Settings.model.rawCollection(),
 });
