+++ package.json
@@ -177,9 +177,9 @@
 		"mkdirp": "^0.5.1",
 		"moleculer": "^0.13.4",
 		"moment": "^2.22.2",
-		"msgpack5": "^4.2.1",
 		"moment-timezone": "^0.5.23",
 		"msgpack5": "^4.2.1",
+		"nats": "^1.0.1",
 		"node-dogstatsd": "^0.0.7",
 		"object-path": "^0.11.4",
 		"pdfjs-dist": "^2.0.550",
